requirements.txt
pillow
django-htmx
streamlit
plotly
django-compressor
django-htmx